+++ app/src/MyComponent.js
@@ -2,8 +2,10 @@ import React from "react";
 import { newContextComponents } from "@drizzle/react-components";
 import logo from "./logo.png";
 
-const { AccountData } = newContextComponents;
-// const { AccountData, ContractData, ContractForm } = newContextComponents;
+// const { AccountData } = newContextComponents;
+const { AccountData, ContractData, ContractForm } = newContextComponents;
+
+const contract = "Minty";
 
 export default ({ drizzle, drizzleState }) => {
   // destructure drizzle and drizzleState from props
@@ -11,9 +13,9 @@ export default ({ drizzle, drizzleState }) => {
     <div className="App">
       <div>
         <img src={logo} alt="drizzle-logo" />
-        <h1>Drizzle Examples</h1>
+        <h1>Minty Fresh</h1>
         <p>
-          Examples of how to get started with Drizzle in various situations.
+          Tool for assisting in NFT uploads and minting.
         </p>
       </div>
 
@@ -26,6 +28,71 @@ export default ({ drizzle, drizzleState }) => {
           units="ether"
           precision={3}
         />
+      </div>
+      
+      <div className="section">
+
+        <h2>{contract}</h2>
+
+        <p>
+          <strong>My Balance: </strong>
+          <ContractData
+            drizzle={drizzle}
+            drizzleState={drizzleState}
+            contract={contract}
+            method="balanceOf"
+            methodArgs={[drizzleState.accounts[0]]}
+          />
+        </p>
+
+        <h3>Mint</h3>
+
+        <p>
+          Mint a new NFT from an existing schema template.
+        </p>
+
+        <p>
+          <ContractForm
+            drizzle={drizzle}
+            drizzleState={drizzleState}
+            contract={contract}
+            method="mintToken"
+            labels={["To Address", "Token URI"]}
+          />
+        </p>
+
+        <h3>Show</h3>
+
+        <p>
+          Get info about an NFT using its token ID.
+        </p>
+
+        <p>
+          <ContractForm
+            drizzle={drizzle}
+            drizzleState={drizzleState}
+            contract={contract}
+            method="tokenURI"
+            labels={["Token ID"]}
+          />
+        </p>
+
+        <h3>Transfer</h3>
+
+        <p>
+          Transfer your NFT to another address using its token ID.
+        </p>
+
+        <p>
+          <ContractForm
+            drizzle={drizzle}
+            drizzleState={drizzleState}
+            contract={contract}
+            method="safeTransferFrom"
+            labels={["From Address", "To Address", "Token ID"]}
+          />
+        </p>
+
       </div>
 
     </div>